LangChain项目/requirements.txt
langchain>=0.3.5
langchain-core>=0.3.8
langchain-community>=0.3.3
langchain-openai>=0.2.2
tiktoken>=0.5.2
python-dotenv>=1.0.0
numpy>=1.24.0
rich>=13.7.0
pydantic>=2.7.0
pypdf>=4.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.24
sentence-transformers>=2.2.2